keras to tensorflow test
